// frontend/src/app/api/model/project-forecasts.ts
/**
 * CloudPulse AI API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ForecastResponse } from './forecast-response';


export interface ProjectForecasts { 
    sdaId: string;
    sdaName: string;
    forecasts: Array<ForecastResponse>;
}